feat: refactor layout by adding Navbar component, implement Footer component, and adjust Menu positioning
