+++ src/components/layout/Footer.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Footer = () => {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default Footer
