+++ src/components/shared/Menu.tsx
@@ -8,7 +8,7 @@ type MenuProps = {
 const Menu = ({ isOpen, toggleMenu }: MenuProps) => {
   return (
     <div
-      className="cursor-pointer absolute top-5 left-5 text-neutral-400 dark:hover:text-neutral-100 hover:text-neutral-500"
+      className="cursor-pointer absolute top-4 left-5 text-neutral-400 dark:hover:text-neutral-100 hover:text-neutral-500"
       onClick={toggleMenu}
     >
       <svg
